requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
Pillow==10.0.1
numpy==1.26.4
opencv-python-headless==4.8.1.86
ultralytics==8.0.196
python-multipart==0.0.6
